agentscope-server/requirements.txt
agentscope
agentscope-runtime
python-dotenv
dashscope>=1.24.4
requests>=2.25.1
psutil>=5.8.0
PyJWT>=2.8.0